refactor: user model defenition feat: turnpoolid in available turns model feat: turnpoolid in desabled turns model migr: remove all migrations and add new inirial migration
