+++ EntityModel/Users/MyUser.cs
@@ -1,0 +1,18 @@
+﻿using EntityModel.Offices;
+using Microsoft.AspNetCore.Identity;
+using System.Diagnostics.CodeAnalysis;
+
+namespace EntityModel.Users
+{
+    public class CostumIdentityUser : IdentityUser
+    {
+        public required string FirstName { get; set; }
+        public required string LastName { get; set; }
+        public bool Status { get; set; }
+
+        //Relations
+
+        public int? OfficeId { get; set; }
+        public Office? Office { get; set; }
+    }
+}
